Add skeleton of Instagram section
Spacing is very terrible and Instagram posts are just images taken from their account. Will need to change to fetch the most recent 4 dynamically. The UWAMS logo is also just a screenshot from the brief, and is not square.

Also have no contact info, and the posts aren't the same size.

Will take more of a look later.